Update outdated API usage

+++ src/java/me/matterz/supernaturals/manager/PriestManager.java
@@ -392,7 +392,7 @@ public class PriestManager extends HumanManager {
 				SuperNManager.sendMessage(snvictim, "You were healed by "
 						+ ChatColor.WHITE + snplayer.getName() + ChatColor.RED
 						+ "!");
-				int health = victim.getHealth()
+				double health = victim.getHealth()
 						+ SNConfigHandler.priestHealAmount;
 				if (health > 20) {
 					health = 20;
